Move card gards out of component module
