+++ projects/storefrontlib/src/lib/checkout/components/multi-step-checkout/multi-step-checkout.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
+import { RouterModule } from '@angular/router';
 import { ShippingAddressModule } from './shipping-address/shipping-address.module';
 import { DeliveryModeModule } from './delivery-mode/delivery-mode.module';
 import { PaymentMethodModule } from './payment-method/payment-method.module';
@@ -8,8 +9,12 @@ import { ReviewSubmitModule } from './review-submit/review-submit.module';
 
 import { MultiStepCheckoutComponent } from './container/multi-step-checkout.component';
 import { CartSharedModule } from '../../../cart/cart-shared/cart-shared.module';
-import { RouterModule } from '@angular/router';
-import { UrlTranslationModule, ConfigModule, CmsConfig } from '@spartacus/core';
+import {
+  UrlTranslationModule,
+  ConfigModule,
+  CmsConfig,
+  CheckoutModule
+} from '@spartacus/core';
 
 @NgModule({
   imports: [
@@ -26,7 +31,8 @@ import { UrlTranslationModule, ConfigModule, CmsConfig } from '@spartacus/core';
       cmsComponents: {
         MultiStepCheckoutComponent: { selector: 'cx-multi-step-checkout' }
       }
-    })
+    }),
+    CheckoutModule
   ],
   declarations: [MultiStepCheckoutComponent],
   entryComponents: [MultiStepCheckoutComponent]
