+++ projects/storefrontlib/src/lib/checkout/components/order-confirmation/order-confirmation.module.ts
@@ -1,6 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
+import { CheckoutModule } from '@spartacus/core';
 import { OrderConfirmationComponent } from './order-confirmation.component';
 
 import { MediaModule } from './../../../ui/components/media/media.module';
@@ -9,7 +10,14 @@ import { CardModule } from '../../../ui/components/card/card.module';
 import { PwaModule } from '../../../pwa/pwa.module';
 
 @NgModule({
-  imports: [CommonModule, MediaModule, CartSharedModule, CardModule, PwaModule],
+  imports: [
+    CommonModule,
+    MediaModule,
+    CartSharedModule,
+    CardModule,
+    PwaModule,
+    CheckoutModule
+  ],
   declarations: [OrderConfirmationComponent],
   exports: [OrderConfirmationComponent]
 })
