+++ projects/storefrontlib/src/lib/ui/pages/guards/guards.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from '@angular/core';
+import { CartNotEmptyGuard } from './cart-not-empty.guard';
+import { RoutingModule, CartModule } from '@spartacus/core';
+
+@NgModule({
+  imports: [RoutingModule, CartModule],
+  providers: [CartNotEmptyGuard]
+})
+export class GardsModule {}
